requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
comm==0.2.2
cryptography==45.0.5
debugpy==1.8.14
decorator==5.2.1
ecdsa==0.19.1
executing==2.2.0
fastapi==0.110.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.6.3
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-telegram-bot==20.8
pyzmq==27.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette==0.36.3
Telethon==1.40.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
uvicorn==0.29.0
wcwidth==0.2.13
yarl==1.20.1
